requirements.txt
django==3.1
djangorestframework==3.12.4
psycopg2==2.9.1
djangorestframework-simplejwt==4.7.2